make hook generic
